feat: Add x2image cog for converting x.com links to images
This commit adds a new cog called x2image that allows users to convert x.com links to images. The cog uses the html2image library to convert the HTML of the tweet into an image. It also uses the wand library to trim the border from the image. The cog provides a command called x2image that takes an x.com link as input and sends the converted image as a file attachment. This feature enhances the utility of the bot by allowing users to easily share images from x.com links.